FEAT: Refactored some code to achieve a faster dynamic risk cost calculation.
The problem now is that the cost for each time step of every rollout is calculated based on the expected locatinos of every future timestep of all obstacles. So the robot will manouevre around all trajectories and cannot find a path where it crosses another trajectory. To solve this, some code has to be changed in the obstacle_class.py file.